services/api/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile 
#
attrs==19.1.0
bcrypt==3.1.6
cffi==1.12.2              # via bcrypt
click==7.0                # via flask
flask-login==0.4.1
flask-pymongo==2.2.0
flask==1.0.2
itsdangerous==1.1.0       # via flask
jinja2==2.10              # via flask
markupsafe==1.1.1         # via jinja2
pycparser==2.19           # via cffi
pymongo==3.7.2            # via flask-pymongo
six==1.12.0               # via bcrypt
werkzeug==0.14.1          # via flask
